middleware/payload: 100% coverage
